manage listings and authorization implemented

+++ app/Http/Controllers/ListingController.php
@@ -5,24 +5,29 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 
-class ListingController extends Controller {
-  public function index() {
+class ListingController extends Controller
+{
+  public function index()
+  {
     return view('listings.index', [
       'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
     ]);
   }
 
-  public function show(Listing $listing) {
+  public function show(Listing $listing)
+  {
     return view('listings.show', [
       'listing' => $listing
     ]);
   }
 
-  public function create() {
+  public function create()
+  {
     return view('listings.create');
   }
 
-  public function store(Request $request) {
+  public function store(Request $request)
+  {
     $formFields = $request->validate([
       'title' => 'required|max:255',
       'company' => 'required|max:255',
@@ -33,7 +38,7 @@ class ListingController extends Controller {
       'description' => 'required|max:5000'
     ]);
 
-    if($request->hasFile('logo')) {
+    if ($request->hasFile('logo')) {
       $formFields['logo'] = $request->file('logo')->store('logos', 'public');
     }
 
@@ -44,13 +49,19 @@ class ListingController extends Controller {
     return redirect('/')->with('message', 'Listing created successfully');
   }
 
-  public function edit(Listing $listing) {
+  public function edit(Listing $listing)
+  {
     return view('listings.edit', [
       'listing' => $listing
     ]);
   }
 
-  public function update(Request $request, Listing $listing) {
+  public function update(Request $request, Listing $listing)
+  {
+    if(auth()->user()->id != $listing->user_id) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $formFields = $request->validate([
       'title' => 'required|max:255',
       'company' => 'required|max:255',
@@ -61,7 +72,7 @@ class ListingController extends Controller {
       'description' => 'required|max:5000'
     ]);
 
-    if($request->hasFile('logo')) {
+    if ($request->hasFile('logo')) {
       $formFields['logo'] = $request->file('logo')->store('logos', 'public');
     }
 
@@ -70,10 +81,21 @@ class ListingController extends Controller {
     return back()->with('message', 'Listing updated successfully');
   }
 
-  public function destroy(Listing $listing) {
+  public function destroy(Listing $listing)
+  {
+    if(auth()->user()->id != $listing->user_id) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $listing->delete();
 
     return back()->with('message', 'Listing deleted successfully');
   }
-}
 
+  public function manage()
+  {
+    return view('listings.manage', [
+      'listings' => auth()->user()->listings()->paginate(10)
+    ]);
+  }
+}
